+++ Reclaimer.Blam/Blam/Halo5/bitmap.cs
@@ -1,4 +1,4 @@
-﻿using Reclaimer.Blam.Common;
+﻿using Reclaimer.Blam.Common.Gen5;
 using Reclaimer.Blam.Utilities;
 using Reclaimer.Drawing;
 using Reclaimer.IO;
