+++ Reclaimer.Blam/Blam/HaloInfinite/Module.cs
@@ -1,4 +1,4 @@
-﻿using Reclaimer.Blam.Common;
+﻿using Reclaimer.Blam.Common.Gen5;
 using Reclaimer.Blam.Utilities;
 using Reclaimer.IO;
 using System.IO;
