+++ Reclaimer/Plugins/CommonModuleViewerPlugin.cs
@@ -1,5 +1,5 @@
 ﻿using Microsoft.Win32;
-using Reclaimer.Blam.Common;
+using Reclaimer.Blam.Common.Gen5;
 using Reclaimer.Blam.HaloInfinite.Oodle;
 using Reclaimer.Blam.Utilities;
 using Reclaimer.Controls.Editors;
