+++ Reclaimer/Plugins/MetaViewer/MetaViewerPlugin.cs
@@ -1,4 +1,5 @@
 ﻿using Reclaimer.Blam.Common;
+using Reclaimer.Blam.Common.Gen5;
 using Reclaimer.Controls.Editors;
 using System.ComponentModel;
 using System.IO;
